Novo controller de favoritos adicionado. NOva view de favoritos adicionada. Usuario consumidor agora pode ver só favoritos. Favoritar e reclamar ainda funciona na tela de favoritos.
