+++ FazendaFeliz.Web/Controllers/FavoritosController.cs
@@ -1,0 +1,53 @@
+﻿using FazendaFeliz.ApplicationCore.Business;
+using FazendaFeliz.ApplicationCore.Interfaces.Repository;
+using FazendaFeliz.ApplicationCore.Interfaces.Service;
+using FazendaFeliz.Web.Models;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
+using System.Diagnostics;
+using Microsoft.AspNetCore.Identity;
+
+namespace FazendaFeliz.Web.Controllers
+{
+    [Authorize]
+    [Route("favoritos")]
+    public class FavoritosController : Controller
+    {
+        public readonly IUsuarioRepository _usuarioRepository;
+        public readonly IAnuncioRepository _anuncioRepository;
+        public readonly IIdentityService _identityService;
+        public readonly IReclamacaoRepository _reclamacaoRepository;
+
+        public FavoritosController(IUsuarioRepository usuarioRepository, IAnuncioRepository anuncioRepository, IReclamacaoRepository reclamacaoRepository, IIdentityService identityService)
+        {
+            _usuarioRepository = usuarioRepository;
+            _anuncioRepository = anuncioRepository;
+            _reclamacaoRepository = reclamacaoRepository;
+            _identityService = identityService;
+        }
+
+        public async Task<IActionResult> Index()
+        {
+            var anuncios = await _anuncioRepository.ObterTodos();         
+            return View("/Views/Favoritos/AnunciosFavoritos.cshtml", anuncios);
+        }
+
+
+        [HttpPost("/favoritos/favoritar")]
+        public async Task<IActionResult> FavoritarAnuncio([FromBody] int idAnuncio)
+        {
+            //OBTER O ANUNCIO DO BANCO E CARREGAR NA PÁGINA
+            var anuncio = await _anuncioRepository.ObterPorId(idAnuncio);
+
+            if (anuncio.Favorito == false)
+                anuncio.Favorito = true;
+            else anuncio.Favorito = false;
+
+            await _anuncioRepository.SaveChanges();
+
+            //return View();
+            return Json(1);
+        }
+    }
+}
