+++ FazendaFeliz.Web/Views/Favoritos/AnunciosFavoritos.cshtml
@@ -9,7 +9,7 @@
     <div class="anuncios-wrapper row">
     @foreach(var anuncio in Model)
     {
-            @if (!anuncio.Oculto)
+            @if (anuncio.Favorito)
             {
                 <section class="col-12 col-lg-4 px-2">
                     <div class="anuncio">
@@ -105,7 +105,6 @@
         }})
         .then(function (response) {
             if(response.status == 200) {
-                toastr.success("Anúncio favoritado com sucesso!");
                 window.location.reload();
             }
             console.log(response);
